docs(UVa): update 490 Rotating Sentences solution and README note

+++ UVa/490_Rotating Sentences/solution.cpp
@@ -13,7 +13,7 @@ int main(){
     }
     for(int i=0;i<maxLen;i++){
         for(int j=(int)sentences.size()-1;j>=0;j--){
-            if(i<sentences[j].length()){
+            if(i<(int)sentences[j].length()){
                 cout<<sentences[j][i];
             }
             else{
